test/meson.build: add mkcomposefs tools to PATH
Some tests run mkcomposefs as a subprocess. If we build composefs as a
subproject (instead of using it from the host system), we need to add
the corresponding tools directory to PATH.

I've not found an explicit function to check if composefs is used as a
subproject or to get the tools path from the meson dependency. So build
the path from the build_root.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>