src/signature: allow bundles with multiple signatures
If we have multiple signers, build the chain for the first. There is
currently no way in RAUC to require more than one and we don't have
use-cases which decide anything based on the details of valid
certificate chains. Accordingly, the additional chains can be ignored
for now. OpenSSL ensures that at least one valid signature exists.

When we support requiring multiple signers, we'll need to extend this
and the bundle info output to support multiple chains.

Sort the list of signer certs so that we have a stable order for
testing.

Signed-off-by: Jan Luebbe <jlu@pengutronix.de>